refactor(apple-wizard): move configure package manager to separate file
